replace old popover component
