+++ libs/core/src/lib/popover/base/base-popover.class.ts
@@ -1,7 +1,6 @@
 import { Directive, ElementRef, EventEmitter, Input, Output } from '@angular/core';
 import { ConnectedPosition, ScrollStrategy } from '@angular/cdk/overlay';
-import { Placement } from '../cdk-overlay/popover-position';
-import { PopoverFillMode } from '../popover-directive/popover.directive';
+import { Placement, PopoverFillMode } from '../popover-position/popover-position';
 
 @Directive()
 export class BasePopoverClass {
