--- libs/core/src/lib/popover/cdk-overlay/cdk-popover.component.html
@@ -1,24 +1,0 @@
-<div
-     *ngIf="!trigger"
-     cdkOverlayOrigin
-     class="fd-popover__control"
-     (keydown)="triggerKeyDownHandler($event)"
-     [ngClass]="additionalTriggerClass">
-    <span class="fd-popover__arrow" [ngClass]="'fd-popover__arrow--' + arrowPosition" *ngIf="!noArrow"></span>
-    <ng-content select="fd-popover-control"></ng-content>
-</div>
-
-<ng-template #templateRef>
-    <div class="fd-popover__popper fd-popover__popper--cdk-custom"
-         [class.fd-popover__popper--no-arrow]="noArrow"
-         [ngClass]="additionalBodyClass"
-         [cdkTrapFocus]="focusTrapped"
-         [cdkTrapFocusAutoCapture]="focusAutoCapture"
-         [style]="marginStyle"
-         (keydown)="bodyKeydownHandler($event)">
-        <ng-content select="fd-popover-body"></ng-content>
-        <ng-content></ng-content>
-    </div>
-</ng-template>
-
-<ng-container #container></ng-container>
