--- libs/core/src/lib/popover/popover-dropdown/popover-dropdown.component.html
@@ -1,15 +1,0 @@
-<div class="fd-dropdown">
-    <button
-        fd-button
-        class="fd-dropdown__control"
-        [fdType]="btnType"
-        [glyph]="glyph"
-        [compact]="compact"
-        [disabled]="disabled"
-        [attr.aria-expanded]="this.disabled ? false : isOpen"
-        [attr.aria-disabled]="this.disabled"
-        aria-haspopup="true"
-    >
-        <ng-content></ng-content>
-    </button>
-</div>
