+++ libs/core/src/lib/popover/popover-position/popover-position.ts
@@ -25,6 +25,7 @@ export const DefaultPositions: ConnectedPosition[] = [
     { originX: 'end', originY: 'top', overlayX: 'end', overlayY: 'bottom' }
 ];
 
+export type PopoverFillMode = 'at-least' | 'equal';
 
 const ARROW_SIZE  = '0.5rem';
