+++ libs/core/src/lib/popover/popover.component.html
@@ -1,35 +1,24 @@
-<div #popoverContainer [ngClass]="addContainerClass ? addContainerClass : ''">
-    <div
-        class="fd-popover__control"
-        [attr.aria-expanded]="this.disabled ? false : isOpen"
-        [attr.aria-disabled]="this.disabled"
-        aria-haspopup="true"
-        (keydown)="handleKeydown($event)"
-        [fdPopover]="popoverBody"
-        [additionalClasses]="additionalClasses"
-        [(isOpen)]="isOpen"
-        (isOpenChange)="openChanged($event)"
-        [noArrow]="noArrow"
-        [disabled]="disabled"
-        [triggers]="triggers"
-        [placement]="placement"
-        [focusTrapped]="focusTrapped"
-        [options]="options"
-        [fillControlMode]="fillControlMode"
-        [closeOnOutsideClick]="closeOnOutsideClick"
-        [closeOnEscapeKey]="closeOnEscapeKey"
-        [appendTo]="appendTo ? appendTo : popoverContainer"
-    >
-        <ng-content select="fd-popover-control"></ng-content>
-        <ng-content select="fd-dropdown-control"></ng-content>
-    </div>
-    <ng-template #popoverBody>
-        <ng-container *ngTemplateOutlet="popoverBodyTpl"></ng-container>
-    </ng-template>
+<div
+     *ngIf="!trigger"
+     cdkOverlayOrigin
+     class="fd-popover__control"
+     (keydown)="triggerKeyDownHandler($event)"
+     [ngClass]="additionalTriggerClass">
+    <span class="fd-popover__arrow" [ngClass]="'fd-popover__arrow--' + arrowPosition" *ngIf="!noArrow"></span>
+    <ng-content select="fd-popover-control"></ng-content>
 </div>
 
-<ng-template #popoverBodyTpl>
-    <div [attr.aria-hidden]="disabled ? true : !isOpen">
+<ng-template #templateRef>
+    <div class="fd-popover__popper fd-popover__popper--cdk-custom"
+         [class.fd-popover__popper--no-arrow]="noArrow"
+         [ngClass]="additionalBodyClass"
+         [cdkTrapFocus]="focusTrapped"
+         [cdkTrapFocusAutoCapture]="focusAutoCapture"
+         [style]="marginStyle"
+         (keydown)="bodyKeydownHandler($event)">
         <ng-content select="fd-popover-body"></ng-content>
+        <ng-content></ng-content>
     </div>
 </ng-template>
+
+<ng-container #container></ng-container>
