+++ libs/core/src/lib/popover/public_api.ts
@@ -1,12 +1,10 @@
 export * from './popover.module';
-export * from './popover.component';
 export * from './popover-body/popover-body.component';
+export * from './popover-directive/popover.directive';
+export * from './popover-directive/popover-container';
 export * from './popover-body/popover-body-directives/popover-body-header.directive';
 export * from './popover-body/popover-body-directives/popover-body-footer.directive';
 export * from './popover-control/popover-control.component';
-export * from './popover-directive/popover-container';
-export * from './popover-directive/popover.directive';
-export * from './popover-dropdown/popover-dropdown.component';
 export * from './popover-control/popover-control.component';
-export * from './cdk-overlay/cdk-popover.component';
-export * from './cdk-overlay/popover-position';
+export * from './popover.component';
+export * from './popover-position/popover-position';
