+++ libs/core/src/lib/shellbar/product-menu/product-menu.component.ts
@@ -2,7 +2,7 @@ import { ChangeDetectionStrategy, Component, Input, ViewChild, ViewEncapsulation
 import { MenuComponent } from '../../menu/menu.component';
 import { ShellbarMenuItem } from '../model/shellbar-menu-item';
 import { Placement } from 'popper.js';
-import { PopoverFillMode } from '../../popover/popover-directive/popover.directive';
+import { PopoverFillMode } from '../../popover/popover-position/popover-position';
 
 /**
  * The component that represents a product menu.
